la norme src

+++ srcs/map_and_color_utils.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   map_utils.c                                        :+:      :+:    :+:   */
+/*   map_and_color_utils.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lpolizzi <lpolizzi@student.42nice.fr>      +#+  +:+       +#+        */
+/*   By: peli <peli@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/11 20:38:37 by lpolizzi          #+#    #+#             */
-/*   Updated: 2025/03/11 20:38:38 by lpolizzi         ###   ########.fr       */
+/*   Updated: 2025/03/12 17:29:12 by peli             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-int	get_rgba(int r, int g, int b, int a)
+unsigned int	get_rgba(int r, int g, int b, int a)
 {
-	return (r << 24 | g << 16 | b << 8 | a);
+	return ((unsigned int)r << 24 | (unsigned int)g << 16 | (unsigned int)b << 8 | a);
 }
 
 int	get_map_value(t_data *data, int x, int y)
